SecureConnect-main/backend/requirements_clean.txt
gunicorn
fastapi
uvicorn
pydantic
python-dotenv
requests
bandit
safety
python-nmap
scapy
python-owasp-zap-v2.4
requests-html
beautifulsoup4
dnspython
shodan
paramiko
detect-secrets
defusedxml
sqlparse
celery
redis
playwright
networkx
pyod
pinecone-client
presidio-analyzer
presidio-anonymizer
pycryptodome
pytest
httpx
python-jose[cryptography]
passlib[bcrypt]
